SummaryTabComponent inside of covid_estimates.js
